api/requirements.txt
fastapi
uvicorn
sqlmodel
passlib
python-jose
psycopg[binary]
